CFOS-102 refactor: Refactor out viewerProbe component
