+++ client/src/components/ViewerTooltip.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import Box from '@mui/material/Box';
 import Paper from '@mui/material/Paper';
 import Typography from '@mui/material/Typography';
