+++ client/src/helpers/probeWidget.js
